+++ .storybook/manager.js
@@ -1,7 +1,5 @@
 import { addons } from "@storybook/addons";
-import { stenaTheme } from "./stena-theme";
 
 addons.setConfig({
   panelPosition: "bottom",
-  theme: stenaTheme
 });