+++ packages/core/src/theme/ThemeContext.tsx
@@ -1,9 +1,0 @@
-import * as React from "react";
-import { defaultTheme } from "./DefaultTheme";
-import { Theme } from "./Theme";
-
-export type ThemeContextValue = Theme;
-
-export const ThemeContext = React.createContext<ThemeContextValue>(
-  defaultTheme
-);