+++ packages/core/src/theme/theme-types/ThemeShadows.ts
@@ -1,5 +1,0 @@
-export interface ThemeShadows {
-  box: string;
-  popover: string;
-  modal: string;
-}