+++ packages/grid-export/src/features/grid-copy-to-clipboard/transformers/__tests__/ConfigTransformer.test.ts
@@ -2,7 +2,7 @@ import {
   testConfig,
   testConfigWithGroups,
   testItems,
-} from "../__mocks__/TestData";
+} from "../__mocks__/CopyToClipboardTestData";
 import { createHtmlConfig } from "../ConfigTransformer";
 
 const fullTable =