+++ packages/panels/src/components/error-panel/ErrorPanel.tsx
@@ -1,5 +1,5 @@
 import { faExclamation } from "@fortawesome/free-solid-svg-icons/faExclamation";
-import { Box, Space, Text, useThemeFields } from "@stenajs-webui/core";
+import { Box, Space, Text } from "@stenajs-webui/core";
 import { Icon } from "@stenajs-webui/elements";
 import * as React from "react";
 import { defaultErrorPanelTheme, ErrorPanelTheme } from "./ErrorPanelTheme";
@@ -13,21 +13,13 @@ export const ErrorPanel: React.FC<ErrorPanelProps> = ({
   text = "Something unexpected happened.",
   theme = defaultErrorPanelTheme,
 }) => {
-  const { colors } = useThemeFields(
-    {
-      colors: {
-        backgroundColor: theme.backgroundColor,
-      },
-    },
-    [theme]
-  );
   return (
     <Box
       width={theme.width}
       height={theme.height}
       justifyContent={"center"}
       alignItems={"center"}
-      background={colors.backgroundColor}
+      background={theme.backgroundColor}
     >
       <Icon icon={faExclamation} size={42} />
       <Space num={4} />